requirements.txt
Flask==3.1.2
pandas==2.3.2
numpy==2.0.2
xlsxwriter==3.2.5
openpyxl==3.1.5
Werkzeug==3.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.10